+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Added
 
+- `Dialog, DialogBase`: Added the `form` and `onSubmit` properties to render a Dialog as a form ([@lorgan3](https://github.com/lorgan3) in [#2074](https://github.com/teamleadercrm/ui/pull/2074))
+
 ### Changed
 
 ### Deprecated